+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Change Log
 
+### 0.0.3
+
+stringId base can be customized in an input box
+
 ### 0.0.2
 
-Description stringId (long & short desc)
+`<name>` and `<description>` stringId (long & short desc)
 
 ### 0.0.1
 
 Initial release.
-